Delete old data and bogus "parallel" data and their cells for Leap comparison
Parallel data is "bogus" in the sense that those workloads weren't really
parallel; I later discovered that OpenMP wouldn't parallelize those. So it's
misleading, and strictly inferior to the other data.

The data we actually included in the NSDI submission was the "multitrial" data,
though in practice there was little variation and so the single-trial data may
still be useful in some cases (though I can't think of any right now).